:sparkles: Add "search filter" block and supporting inner blocks
The search filter block renders a search form and one or more
sections of variable content to be searched. It uses innerBlocks
and depends on two child blocks: the search filter input block to
render the search form, and the search filter section block to
contain the variable content to be searched and filtered.