:milky_way: Assign aria-live to dynamic search region, fix #2
The standard "aria-live" attribute seemed inappropriate for this
feature during testing. It announced too much to be particularly
useful, especially for long sections of searchable content.
Nevertheless, we've set the attribute on the searchable content
container to indicate that it is a live region, but set the
value to "off" to prevent noise.