:bug: Fix #18 load styles on all archive pages
HRS Blocks styles were not loading on archive pages because of the
conditional `has_block()` check before enqueueing styles. The
`has_blocks` function does not work on archive page content because
the string it searches for does not exist. To avoid this, we will
instead load the styles on all non-singular pages, since the styles
will probably be needed anyway and can be cached.