Fix #67 use client-side wp.data method to hide blocks
Had been using the `allowed_blocks` hook server side, but there is
no way to get a reliable list of all registered blocks to filter
against. Switch to using the `hideBlockTypes` function in the
wp.data.dispatch( 'core/edit-post' ) store. This adds the block
manually to the "hide block" preferences on the client side.